std: Move platform specific path code into sys
