+++ src/libstd/sys/unix/mod.rs
@@ -41,6 +41,7 @@ pub mod mutex;
 pub mod net;
 pub mod os;
 pub mod os_str;
+pub mod path;
 pub mod pipe;
 pub mod process;
 pub mod rand;
