+++ src/libstd/sys/windows/mod.rs
@@ -29,6 +29,7 @@ pub mod mutex;
 pub mod net;
 pub mod os;
 pub mod os_str;
+pub mod path;
 pub mod pipe;
 pub mod process;
 pub mod rand;
